feat(Dependencies): allow configuration of the dependency rendering in the App
