+++ packages/admin/cms-admin/src/dam/config/DamConfigContext.tsx
@@ -1,7 +1,20 @@
+import { ApolloClient } from "@apollo/client";
 import * as React from "react";
+
+export interface DamDependencyRenderInfo {
+    type: React.ReactNode;
+    name: React.ReactNode;
+    secondaryInfo: React.ReactNode;
+}
 
 export interface DamConfig {
     additionalMimeTypes?: string[];
+    dependencyRenderInfoProvider?: {
+        [key: string]: {
+            getRenderInfo: (id: string, apolloClient: ApolloClient<unknown>) => Promise<DamDependencyRenderInfo> | DamDependencyRenderInfo;
+            renderCustomContent?: (renderInfo: DamDependencyRenderInfo) => React.ReactNode;
+        };
+    };
 }
 
 export const DamConfigContext = React.createContext<DamConfig | undefined>(undefined);
